requirement.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.9
django-summernote==0.8.20.0
django-taggit==5.0.1
jdatetime==4.1.1
persian-tools==0.0.11
persiantools==3.0.1
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
webencodings==0.5.1
